requirements.txt
# Core dependencies
requests>=2.31.0
configparser>=5.3.0

# Google API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0

# Document processing dependencies
PyPDF2>=3.0.1
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
pypandoc>=1.11
pdf2image>=1.16.3

# Data processing
pandas>=2.0.0
